feat: create email
To create an email it's necessary to have a credential. So, we will get the credential from the cookie 'accessToken'. The payload of that cookie is the id of the registered credential.
Once we got that cookie we can register the email making the relation with the credential.
Email model will set every new email's status to 'pending' as default.